requirements.txt
fastapi
uvicorn[standard]
adafruit-circuitpython-neopixel-spi